added 400 errors for needed cases

+++ src/app/internal/data/repositories/user.py
@@ -7,7 +7,7 @@ from app.internal.domain.interfaces.user import IUserRepository
 
 
 class UserRepository(IUserRepository):
-    def create(self, name: str, token_encrypted: str, token_hash: str) -> dict:
+    def create(self, name: str, token_encrypted: str, token_hash: str) -> dict | None:
         user, created = User.objects.get_or_create(
             name=name,
             defaults={
@@ -19,12 +19,15 @@ class UserRepository(IUserRepository):
             Balance.objects.bulk_create(
                 [Balance(user=user, tool_id=tool_id, amount=0) for tool_id in Tool.objects.values_list('id', flat=True)]
             )
-        return {'id': user.id, 'name': user.name, 'role': user.role, 'api_key_encrypted': user.token_encrypted}
+            return {'id': user.id, 'name': user.name, 'role': user.role, 'api_key_encrypted': user.token_encrypted}
+        return
 
     def get_user_by_token_hash(self, token_hash: str) -> dict | None:
         return User.objects.filter(token_hash=token_hash).values_list('id', 'role').first()
 
     def delete_user_by_id(self, user_id: UUID) -> dict | None:
         user_data = User.objects.filter(id=user_id).values('id', 'name', 'role', 'token_encrypted').first()
-        User.objects.filter(id=user_id).delete()
-        return user_data
+        if user_data:
+            User.objects.filter(id=user_id).delete()
+            return user_data
+        return
